requirements.txt
Django==2.1.10
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.1
urllib3==1.26.4
